pyproject.toml
[project]
name = "auto-rewrite-commit"
version = "0.1.0"
description = "AI-powered Git history rewriter"
requires-python = ">=3.12,<3.13"
dependencies = [
    "gitpython>=3.1.40",
    "typer>=0.9.0",
    "pyyaml>=6.0",
    "openai>=1.0.0",
    "rich>=13.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
git-ai-rewrite = "src.main:app"
